feat(gameboard): Setup gameboard with border nodes
When the gameboard is initialized it is now done
so with the nodes on the borders having their
edges set appropriately to null where an edge
doesn't exist. For example, the top right corner
node 99 doesn't have an edge to the north or to
the east, therefore it's edges for north and east
will be set to null.