feat(computer logic): Enhance computer difficulty
The computer now has a range of different methods
to make different moves throughout the game based
on the conditions. After 20 moves have been made
the computer will have a 50% chance to hit a ship
on it's turn. Before this, if the computer has
hit a ship twice it will continue to attack that
ship. This makes the computer a more worthy
adversary. I still might tweak this so that it
feels fair to the player, though ultimately games
that are close are the most fun, but I still must
make the computer have a human touch to it's
playstyle.